[kde.eclass] Patch for >=autoconf-2.65.
This modification to the eclass causes its kde_src_configure function to try
applying a patch to admin/acinclude.m4.in if that file exists.  The patch is
intended to improve compatibility with recent versions of autotools.

If the patch fails, a warning is reported but the ebuild continues.  This is
intended for cases where a file of the same name but with different content
exists inside the package tarball.

The patch is provided as a heredoc because many kde packages will need it,
it is difficult to know which ones do, and adding it to SRC_URI feels like
overkill in any case.  The patch is set to ignore whitespace changes, as one
input line has a trailing space, but the Gentoo eclass mode for Emacs strips
trailing spaces from the heredoc.

References:
http://www.gnu.org/software/autoconf/manual/html_node/Expanded-Before-Required.html
http://bugs.gentoo.org/279482
http://websvn.kde.org/branches/KDE/3.5/kde-common/admin/acinclude.m4.in