[kde.eclass] Don't call base_src_unpack.
As base_src_unpack will unconditionally call base_src_prepare for EAPI<2,
and as we are collecting the full list of patches only after unpacking and
therefore have to call base_src_prepare later on, we simply cannot call
base_src_unpack else patches would get applied twice.

This was not a problem back when base.eclass used stages, because we could
execute the unpack stage of base_src_unpack without running the auto-patch
stage as well.  Those days are gone, though, so we need to cope.