[kde-meta.eclass] Separate unpack and prepare phases.
This ensures the well-established call order even for EAPI 2.  kde_prepare
and the patches it applies via base_prepare are part of the prepare phase of
EAPI 2, while the makefiles stage of kde-meta_src_unpack would be executed
at the end of the unpack phase, i.e. before patches got applied.  This is
in conflict with the procedure recommended by the kde-meta_src_unpack
documentation comment, which asks for the makefiles stage to be executed
after patches have been applied.

The new setup adds a src_prepare phase to kde-meta as well. It will delegate
to kde_src_prepare for EAPI 2 and later, but not for EAPI 0 and 1, as in
these cases, kde_src_unpack called by kde-meta_src_unpack already invoked
kde_src_prepare.

While the multiple stages approach in kde.eclass has been removed, it is
retained in kde-meta.eclass for the time being, as there are a few ebuilds
which rely on it.  They should be ported to EAPI 2 first.